Feature/74 (#85)
* using preferred font construction

* fixed warning on twiddle usage

* using latest projucer for bin scramlber

* added no warnings to build

* added cpm

* fixed all warnings

* updated gitignore

+++ BinScrambler/Source/BinScramblerFFTProcessor.cpp
@@ -6,8 +6,8 @@ void BinScramblerFFTProcessor::spectral_process(const PolarVector &in, PolarVect
     const int phase = m_phase;
     const int maxPhase = m_maxPhase;
 
-    for(int i=0; i<bins; ++i){
-        out[i].m_mag = utilities::interp_lin<FftDecimal>(in[m_AInd->at(i)].m_mag, in[m_BInd->at(i)].m_mag, (FftDecimal)phase/(FftDecimal)maxPhase);
-        out[i].m_phase = utilities::interp_lin<FftDecimal>(in[m_AInd->at(i)].m_phase, in[m_BInd->at(i)].m_phase, (FftDecimal)phase/(FftDecimal)maxPhase);
+    for(size_t i=0; i<(size_t)bins; ++i){
+        out[i].m_mag = utilities::interp_lin<FftDecimal>(in[(size_t)m_AInd->at(i)].m_mag, in[(size_t)m_BInd->at(i)].m_mag, (FftDecimal)phase/(FftDecimal)maxPhase);
+        out[i].m_phase = utilities::interp_lin<FftDecimal>(in[(size_t)m_AInd->at(i)].m_phase, in[(size_t)m_BInd->at(i)].m_phase, (FftDecimal)phase/(FftDecimal)maxPhase);
     }
 }
